Add NoteSchemaPerson
This model includes the related Person in the Schema. You need a new Schema to prevent a Recursion error because the PeopleSchema contains all notes by using the NotesSchema. If you’d add `people` to the NotesSchema, then you’d include the People, which have Notes, which have People, and so on …